Component changes and decreasing the size of the board
Changed from using the arduino micro to the now using the ATMEGA328P
microchip. Also changed the the the Max232D so that we can run of 5
volts and we no longer have 3.3V on the board. The voltage display and
the push button have now been moved to the back so that we can make the
board smaller. The transisors were replaced with Mosfets and a crystal
and programming pins were added. The board is now also receiving its
power, and data from bsm via serial line.