Reroute and label
Minor changes made to the positioning of the capacitors. Some traces
were changed due to the capacitor changes and some changes were made to
keep the trace from jumping between levels. Labels were positioned on
the board and MRDT symbol was placed on the board as well as a brief
board discription.